--- ui/src/components/Appliances/Devices/Devices.vue
@@ -1,9 +1,0 @@
-<template>
-  <DevicesTable />
-</template>
-
-<script lang="ts" setup>
-import DevicesTable from './DevicesTable.vue'
-
-</script>
-<style></style>
