--- ui/src/components/Appliances/Devices/mock.ts
@@ -1,1 +1,0 @@
-// could be used for json-server package
