+++ ui/src/router/index.ts
@@ -1,6 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Dashboard from '@/containers/Dashboard.vue'
-import Appliances from '@/containers/Appliances.vue'
 
 const router = createRouter({
   history: createWebHistory(),
@@ -13,12 +12,7 @@ const router = createRouter({
     {
       path: '/appliances',
       name: 'Appliances',
-      component: Appliances
-    },
-    {
-      path: '/devices',
-      name: 'Devices',
-      component: () => import('@/components/Appliances/Devices/Devices.vue')
+      component: ()=>import('@/containers/Appliances.vue')
     },
     {
       path: '/:pathMatch(.*)*', // catch other paths and redirect
